Update README.md
In creating a New Nginx Configuration File the the name of file n8n without extension file but at the point 3. Enable the Configuration sudo ln -s call the file n8n.conf is have an extension file so I give extension too when Creating a New Nginx at the point 1. an I repair the nano the program can't read // as a command so i change it to # may be it help fully.